Builder#formatで行っていた処理を、他のクラスに切り出しました
