+++ 08.ls_object/lib/list_command.rb
@@ -14,7 +14,6 @@ class ListCommand
 
   def list_paths
     @builder.collect_path
-    @builder.format
     @builder.result
   end
 end
